chore(persistence): refactor batch persistence classes and tests for clarity
